+++ account/nuls-account/src/test/java/io/nuls/account/service/AccountServiceTest.java
@@ -39,7 +39,7 @@ import static org.junit.Assert.*;
 public class AccountServiceTest {
 
     protected static AccountService accountService;
-    protected short chainId = 12345;
+    protected int chainId = 12345;
     protected String password = "a12345678";
 
     @BeforeClass
